~ Refactored classes + Added new page for set flow + Added new Main activity

+++ app/src/main/java/io/github/michael_bailey/gym_log_book/data_type/ExerciseItem.kt
@@ -8,7 +8,7 @@ data class ExerciseItem(
 	override val id: Long,
 	val date: LocalDate,
 	var exercise: String,
-	var setNumber: Long,
-	var weight: Long,
-	var reps: Long,
+	var setNumber: Int,
+	var weight: Int,
+	var reps: Int,
 ) : Identifiable()
